reorganised d05

+++ d01/ex04.py
@@ -1,7 +1,7 @@
 num1 = "50"
 num2 = "5"
 
-num1 = int(num1)
-num2 = int(num2)
+num1 = float(num1)
+num2 = float(num2)
 
 print(num1/num2)
